decimal number rounded

+++ app/src/main/java/com/example/outlaycalc/AddMovementActivity.kt
@@ -33,7 +33,7 @@ class AddMovementActivity : AppCompatActivity() {
 
     fun createMovement() {
         var newMovement = Movement()
-        newMovement.amount = inputTxtAmount.text.toString().toDouble()
+        newMovement.amount = inputTxtAmount.text.toString().toFloat()
         newMovement.description = inputTxtDescription.text.toString()
         if (radioIngress.isChecked) {
             newMovement.outlay = false
